+++ instrumentation/apache-httpclient/apache-httpclient-2.0/.telemetry/scope.yaml
@@ -1,0 +1,5 @@
+scope:
+  name: io.opentelemetry.apache-httpclient-2.0
+  version: 2.14.0-alpha-SNAPSHOT
+  schemaUrl: null
+  attributes:
