+++ instrumentation/apache-httpclient/apache-httpclient-2.0/javaagent/build.gradle.kts
@@ -19,5 +19,6 @@ dependencies {
 tasks {
   withType<Test>().configureEach {
     systemProperty("testLatestDeps", findProperty("testLatestDeps") as Boolean)
+    systemProperty("collectMetadata", findProperty("collectMetadata")?.toString() ?: "false")
   }
 }
