+++ instrumentation/clickhouse-client-0.5/javaagent/build.gradle.kts
@@ -29,6 +29,7 @@ tasks {
 
   val testStableSemconv by registering(Test::class) {
     jvmArgs("-Dotel.semconv-stability.opt-in=database")
+    systemProperty("collectMetadata", findProperty("collectMetadata")?.toString() ?: "false")
   }
 
   check {
